added implements_trait function to environment
